+++ betterrandom/src/main/java/io/github/pr0methean/betterrandom/prng/ThreadLocalRandomWrapper.java
@@ -3,7 +3,6 @@ package io.github.pr0methean.betterrandom.prng;
 import static io.github.pr0methean.betterrandom.util.BinaryUtils.convertBytesToLong;
 
 import com.google.common.base.MoreObjects.ToStringHelper;
-import io.github.pr0methean.betterrandom.seed.RandomSeederThread;
 import io.github.pr0methean.betterrandom.seed.SeedException;
 import io.github.pr0methean.betterrandom.seed.SeedGenerator;
 import io.github.pr0methean.betterrandom.util.Java8Constants;
@@ -159,15 +158,6 @@ public class ThreadLocalRandomWrapper extends RandomWrapper {
     return getWrapped().nextGaussian();
   }
 
-  /**
-   * Not supported, because this class uses a thread-local seed.
-   * @param thread ignored.
-   * @throws UnsupportedOperationException always.
-   */
-  @Override public void setSeederThread(@Nullable final RandomSeederThread thread) {
-    throw new UnsupportedOperationException("This can't be reseeded by a RandomSeederThread");
-  }
-
   @Override protected boolean useParallelStreams() {
     return true;
   }
@@ -191,8 +181,8 @@ public class ThreadLocalRandomWrapper extends RandomWrapper {
     }
   }
 
-  @Override protected void recordEntropySpent(final long bits) {
-    getWrapped().recordEntropySpent(bits);
+  @Override protected void debitEntropy(final long bits) {
+    getWrapped().debitEntropy(bits);
   }
 
   @Override public long getEntropyBits() {